pyproject.toml
[project]
name = "tiny-knn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "fastapi>=0.116.1",
    "fastembed>=0.7.2",
    "numpy>=2.3.2",
    "pillow>=11.0.0",
    "requests>=2.32.5",
    "torch>=2.2.0",
    "tqdm>=4.67.1",
    "transformers>=4.40.0",
    "uvicorn>=0.35.0",
]
